+++ vscode/src/testutils/emptyDisposable.ts
@@ -1,0 +1,3 @@
+import { Disposable } from './Disposable'
+
+export const emptyDisposable = new Disposable(() => {})